+++ app/components/gh-search-input-trigger.js
