+++ app/components/modal-markdown-help.js
@@ -1,0 +1,4 @@
+import ModalComponent from 'ghost-admin/components/modal-base';
+
+export default ModalComponent.extend({
+});
