+++ app/components/modal-suspend-user.js
@@ -1,11 +1,13 @@
-import ModalComponent from 'ghost-admin/components/modals/base';
+import ModalComponent from 'ghost-admin/components/modal-base';
+import {alias} from '@ember/object/computed';
 import {invokeAction} from 'ember-invoke-action';
 import {task} from 'ember-concurrency';
 
 export default ModalComponent.extend({
-    user: null,
 
-    transferOwnership: task(function* () {
+    user: alias('model'),
+
+    suspendUser: task(function* () {
         try {
             yield invokeAction(this, 'confirm');
         } finally {
@@ -15,7 +17,7 @@ export default ModalComponent.extend({
 
     actions: {
         confirm() {
-            this.get('transferOwnership').perform();
+            return this.get('suspendUser').perform();
         }
     }
 });
