+++ app/components/modal-unsuspend-user.js
@@ -1,18 +1,13 @@
-import ModalComponent from 'ghost-admin/components/modals/base';
+import ModalComponent from 'ghost-admin/components/modal-base';
 import {alias} from '@ember/object/computed';
-import {computed} from '@ember/object';
 import {invokeAction} from 'ember-invoke-action';
 import {task} from 'ember-concurrency';
 
 export default ModalComponent.extend({
 
-    tag: alias('model'),
+    user: alias('model'),
 
-    postInflection: computed('tag.count.posts', function () {
-        return this.get('tag.count.posts') > 1 ? 'posts' : 'post';
-    }),
-
-    deleteTag: task(function* () {
+    unsuspendUser: task(function* () {
         try {
             yield invokeAction(this, 'confirm');
         } finally {
@@ -22,7 +17,7 @@ export default ModalComponent.extend({
 
     actions: {
         confirm() {
-            this.get('deleteTag').perform();
+            return this.get('unsuspendUser').perform();
         }
     }
 });
