--- app/components/modals/markdown-help.js
@@ -1,4 +1,0 @@
-import ModalComponent from 'ghost-admin/components/modals/base';
-
-export default ModalComponent.extend({
-});
