+++ app/templates/components/gh-search-input-trigger.hbs
