+++ app/templates/components/gh-search-input.hbs
@@ -5,7 +5,7 @@
     onopen=(action "onFocus")
     onclose=(action "onBlur")
     searchEnabled=false
-    triggerComponent="gh-search-input/trigger"
+    triggerComponent="gh-search-input-trigger"
     renderInPlace=true
     loadingMessage="Loading"
     as |name select|}}
