+++ app/templates/components/modal-delete-all.hbs
