+++ app/templates/components/modal-delete-subscriber.hbs
