+++ app/templates/components/modal-delete-theme.hbs
