+++ app/templates/components/modal-re-authenticate.hbs
