+++ app/templates/components/modal-theme-warnings.hbs
