+++ app/templates/components/modal-upload-image.hbs
