+++ tests/integration/components/modals/import-subscribers-test.js
@@ -4,8 +4,8 @@ import {describe, it} from 'mocha';
 import {expect} from 'chai';
 import {setupComponentTest} from 'ember-mocha';
 
-describe('Integration: Component: modals/import-subscribers', function() {
-    setupComponentTest('modals/import-subscribers', {
+describe('Integration: Component: modal-import-subscribers', function() {
+    setupComponentTest('modal-import-subscribers', {
         integration: true
     });
 
@@ -14,12 +14,12 @@ describe('Integration: Component: modals/import-subscribers', function() {
         // Handle any actions with this.on('myAction', function(val) { ... });
         // Template block usage:
         // this.render(hbs`
-        //   {{#modals/import-subscribers}}
+        //   {{#modal-import-subscribers}}
         //     template content
-        //   {{/modals/import-subscribers}}
+        //   {{/modal-import-subscribers}}
         // `);
 
-        this.render(hbs`{{modals/import-subscribers}}`);
+        this.render(hbs`{{modal-import-subscribers}}`);
         expect(this.$()).to.have.length(1);
     });
 });
